rename: provider class들에 대해 module로 패키징
