+++ src/main/java/com/github/bluekey/processor/MusicDistributorType.java
@@ -1,5 +1,6 @@
 package com.github.bluekey.processor;
 
+import com.github.bluekey.processor.provider.AtoDistributorExcelFileProvider;
 import lombok.Getter;
 
 @Getter
